Bezier Lines
Replaced the way the terrain is rendered from rectangles to Bezier Lines. Also, redesigned the way the map is looped: now all the obstacle objects are shifted to the front in a for-loop, and the last element is re-initialized. Lastly, the way the obstacle components are calculated was moved to the separate function: 'obstacleInitData'.