Resources and particles
I've added Fredoka font to the project; Now the text fits the art-style. What's more, I've started working on the particle system for my game: it's in the pretty early stage,, but for now it works fine. I want to make it more expandable, reusable, etc.; add it to collectibles, add an option for sprite varieties, etc. First commit in a long time, and it's also the first commit to the GitLab (for this project and in my life in general, lol).